feat: add the comments on the photo

+++ frontend/src/components/Photo/Photo.tsx
@@ -1,17 +1,26 @@
 import styles from "./Photo.module.css";
-import { Link, useNavigate } from "react-router-dom";
+import { Link, useSearchParams } from "react-router-dom";
 import { uploads } from "../../utils/config";
 import { FaRegComment, FaRegHeart, FaHeart } from "react-icons/fa";
+import { RiChatDeleteLine } from "react-icons/ri";
 import { BsPencilSquare, BsTrash } from "react-icons/bs";
 import { IconContext } from "react-icons/lib";
-import { IData, IPhoto } from "../../utils/type";
+import { IComment, IData, IPhoto } from "../../utils/type";
 import { getTimeStamp } from "../../utils/getTimeStamp";
 import { useSelector } from "react-redux";
 import { AppDispatch, RootState } from "../../store";
 import { useDispatch } from "react-redux";
-import { deletePhoto, getAllPhotos, likePhoto } from "../../slices/photoSlice";
-import { useEffect } from "react";
+import {
+  commentPhoto,
+  deleteComment,
+  deletePhoto,
+  getAllPhotos,
+  getPhotoById,
+  likePhoto,
+} from "../../slices/photoSlice";
+import { FormEvent, useEffect, useState } from "react";
 import { useResetMessage } from "../../hooks/useResetMessage";
+import { getAllUsers } from "../../slices/usersSlices";
 
 interface IPhotoProps {
   photo: IPhoto;
@@ -21,14 +30,21 @@ interface IPhotoProps {
 
 export function Photo({ photo, user, comments }: IPhotoProps) {
   const { user: userAuth } = useSelector((state: RootState) => state.auth);
+  const { users } = useSelector((state: RootState) => state.users);
   const dispatch = useDispatch<AppDispatch>();
-  const navigate = useNavigate();
   const resetMessage = useResetMessage(dispatch);
+  const [params, setParams] = useSearchParams();
+  const [caption, setCaption] = useState<string>("");
 
   const handleDelete = (id: string) => {
-    dispatch(deletePhoto(id));
-    navigate("/");
-    resetMessage();
+    let alert = confirm("Deseja mesmo apagar a foto?");
+    if (alert) {
+      dispatch(deletePhoto(id));
+      params.delete("photo");
+      setParams(params);
+      resetMessage();
+      dispatch(getAllPhotos());
+    }
   };
 
   const handleLike = (photo: IPhoto) => {
@@ -36,13 +52,45 @@ export function Photo({ photo, user, comments }: IPhotoProps) {
     resetMessage();
   };
 
+  const handleDeleteComment = (photo: IPhoto, comment: IComment) => {
+    const ids = {
+      idPhoto: photo._id,
+      idComment: comment._id,
+    };
+
+    let alert = confirm(
+      `Deseja mesmo apagar o comentário: ${comment.comment}?`
+    );
+    if (alert) {
+      resetMessage();
+      dispatch(deleteComment(ids));
+      dispatch(getPhotoById(photo._id as string));
+    }
+  };
+
+  const handleComment = (e: FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+
+    const commentData = {
+      comment: caption,
+      _id: photo._id,
+    };
+
+    dispatch(commentPhoto(commentData));
+    setCaption("");
+    resetMessage();
+    dispatch(getPhotoById(photo._id as string));
+  };
+
   useEffect(() => {
-    dispatch(getAllPhotos);
+    dispatch(getPhotoById(photo._id as string));
+    dispatch(getAllPhotos());
+    dispatch(getAllUsers());
   }, [dispatch]);
 
   return (
-    <article key={photo._id}>
-      <div>
+    <article key={photo?._id}>
+      <div style={{ maxWidth: "468px" }}>
         <header className={styles.headerPhoto}>
           <Link to={`/${user?.username}`}>
             <div className={styles.userPostPhoto}>
@@ -69,20 +117,25 @@ export function Photo({ photo, user, comments }: IPhotoProps) {
               className={styles.likes}
               onClick={() => handleLike(photo as IPhoto)}
               role="button"
+              tabIndex={0}
             >
               {photo.likes && photo.likes?.includes(userAuth!._id!) ? (
                 <IconContext.Provider value={{ color: "red", size: "24" }}>
                   <FaHeart />
                 </IconContext.Provider>
               ) : (
-                <IconContext.Provider value={{ color: "white", size: "24" }}>
+                <IconContext.Provider
+                  value={{ className: `${styles.icons}`, size: "24" }}
+                >
                   <FaRegHeart />
                 </IconContext.Provider>
               )}
               <span>{photo.likes?.length}</span>
             </div>
             <div className={styles.comments}>
-              <IconContext.Provider value={{ color: "white", size: "24" }}>
+              <IconContext.Provider
+                value={{ className: `${styles.icons}`, size: "24" }}
+              >
                 <FaRegComment />
               </IconContext.Provider>
               <span>{photo.comments?.length}</span>
@@ -90,9 +143,15 @@ export function Photo({ photo, user, comments }: IPhotoProps) {
           </div>
           {userAuth?._id === photo.username && comments && (
             <div className={styles.edit}>
-              <IconContext.Provider value={{ color: "white", size: "24" }}>
+              <IconContext.Provider
+                value={{ className: `${styles.icons}`, size: "24" }}
+              >
                 <BsPencilSquare />
-                <BsTrash onClick={() => handleDelete(photo._id as string)} />
+                <BsTrash
+                  onClick={() => handleDelete(photo._id as string)}
+                  role="button"
+                  tabIndex={0}
+                />
               </IconContext.Provider>
             </div>
           )}
@@ -103,6 +162,84 @@ export function Photo({ photo, user, comments }: IPhotoProps) {
           </Link>
           <h3>{photo.title}</h3>
         </div>
+        {comments && (
+          <div className={styles.commentsWrapper}>
+            <form onSubmit={handleComment} className={styles.textAreaAndButton}>
+              <textarea
+                aria-label="Adicione um comentário..."
+                placeholder="Adicione um comentário..."
+                autoComplete="off"
+                autoCorrect="off"
+                className={styles.textArea}
+                onChange={(e) => setCaption(e.target.value)}
+                value={caption}
+                style={caption.length >= 62 ? { height: "80px" } : {}}
+              />
+              <button disabled={caption.length <= 0}>Publicar</button>
+            </form>
+
+            {photo.comments && photo.comments?.length > 0 ? (
+              <ul className={styles.ul}>
+                {photo?.comments?.map((comment: IComment) =>
+                  users
+                    ?.filter((user: IData) => user._id === comment.username)
+                    .map((filteredUser) => (
+                      <li key={filteredUser._id}>
+                        <Link
+                          to={`/${filteredUser.username}`}
+                          className={styles.imageUserComment}
+                        >
+                          <img
+                            src={`${uploads}/users/${filteredUser?.profileImage}`}
+                            alt={`Foto de ${filteredUser?.username}`}
+                          />
+                        </Link>
+                        <div className={styles.commentAndTime}>
+                          <div className={styles.commentAndEdit}>
+                            <div className={styles.commentAndusername}>
+                              <Link to={`/${filteredUser.username}`}>
+                                {filteredUser.username}
+                              </Link>
+                              <span>{comment.comment}</span>
+                            </div>
+                            {comment.username === userAuth?._id && (
+                              <div
+                                className={styles.delete}
+                                role="button"
+                                aria-label="Apagar comentário"
+                                onClick={() =>
+                                  handleDeleteComment(
+                                    photo as IPhoto,
+                                    comment as IComment
+                                  )
+                                }
+                              >
+                                <IconContext.Provider
+                                  value={{ color: "#fe6464", size: "18" }}
+                                >
+                                  <RiChatDeleteLine />
+                                </IconContext.Provider>
+                              </div>
+                            )}
+                          </div>
+                          <div className={styles.timeComment}>
+                            <span>
+                              {getTimeStamp(comment.createdAt as string)}
+                            </span>
+                          </div>
+                        </div>
+                      </li>
+                    ))
+                )}
+              </ul>
+            ) : (
+              <span className={styles.notComment}>
+                Ainda não há nenhum comentário.
+                <p>Seja o primeiro a comentar</p>
+              </span>
+            )}
+          </div>
+        )}
       </div>
     </article>
   );
